requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
cyrtranslit==1.0
Django==3.2
django-environ==0.8.1
django-jazzmin==2.4.9
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.3
openpyxl==3.0.9
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
